requirements.txt
torch==2.0.1
transformers>=4.38.2
timm>=0.9.16
accelerate
sentencepiece
attrdict
einops
hydra-core>=1.3.2
shortuuid>=1.0.13
rich>=13.8.1
opencv-python>=4.10.0.84
thefuzz>=0.22.1
ruff>=0.6.5
fastapi>=0.115.0
transformers>=4.44.2
openai>=1.46.0
av>=13.1.0
uvicorn==0.32.0
attrs
gym3